Add files via upload
The Calculator App is a simple, lightweight web application that performs basic arithmetic computations. It lets users enter numbers and perform arithmetic operations such as addition, subtraction, multiplication, and division. The software has an intuitive UI, making it simple to use. It was designed with accessibility in mind and has large, clearly labeled buttons for each digit and operation. This project was developed to provide an introduction to front-end web programming, leveraging essential web technologies such as HTML, CSS, and JavaScript. Building this app allowed me to reaffirm my comprehension of the following concepts:

- Event handling in JavaScript entails capturing user input (button clicks) and changing the display.
- DOM manipulation involves dynamically updating the calculator screen with the current input, results, and operations.
- CSS layout approaches include creating a responsive design that works on desktop and ensures the app looks fantastic across all screen sizes.

Why This Project:
This project was built as a fun and practical way to improve my skills in front-end development while creating a useful tool. This project was aimed at improving my front-end development abilities and providing a helpful tool. It also demonstrated how to build a simple JavaScript application with a clean, user-friendly UI.